add work around
there was a bug where if price was a string we could not compare it to an int. fixed it so that if it is a string it converts to an int and it if cannot be converted then it skipps it. so it if it is none or 'a' it will raaise a value error or type error and skip that part