bug fix
fixed a bug that caused on class 'き' to not be recognized this was a problem with trying to match the name of the class with the data it self and not the file name.  on top of that there were decoding issues where I would not properly check the name of the files becasue they were unzipped and didnt properly decode. by encoding and decoding again we end up with the correct file names that we can use to check the correct class name for the correct file